Factor out recentIssuesHandler
